Added SIMD implementation of FWT.
It is currently only limited to Pa=3, but extension to Pa=5 should be
possible via the use of some arrays of __m256{,i} values. This will mean
changing the operations that are run, and including some looping over
this array.

Can't currently compare the results as the gpu code doesn't work for
Na!=32, and not sure what's wrong with the CPU code - that should be
working for any Pa.

Looks to be around 8.3x speed-up for CPU+SIMD vs CPU.